refactor(agglayer): return AccountBuilder from the account builder helpers
Address review feedback on the deployment builder API:
- drop the AggLayerBridgeAccountBuilder and AggLayerFaucetAccountBuilder
  wrappers; AggLayerBridge::account_builder and AggLayerFaucet::account_builder
  now return the underlying AccountBuilder directly
- move the test-only helpers (zero_fee_policy_manager, faucet_account_builder,
  create_existing_agglayer_faucet) into miden_agglayer::testing
