+++ crates/miden-testing/tests/agglayer/faucet_helpers.rs
@@ -1,6 +1,7 @@
 extern crate alloc;
 
-use miden_agglayer::{AggLayerFaucet, create_existing_agglayer_faucet};
+use miden_agglayer::AggLayerFaucet;
+use miden_agglayer::testing::create_existing_agglayer_faucet;
 use miden_protocol::Felt;
 use miden_protocol::account::auth::AuthScheme;
 use miden_protocol::asset::FungibleAsset;
